[Container Scalability] Basic implementation of MySqlAccountService  (#1624)
Basic implementation of MySqlAccountService with following functionalities.
a) Fetch all accounts and container metadata from DB and initialize cache on boot-up,
b) Implement AccountService interface update APIs to write changes to mysql database and getter APIs to read changes from cache
c) Scheduler thread to sync changes from DB
d) Add MySqlAccountServiceFactory and MySqlAccountServiceConfig

+++ ambry-tools/src/main/java/com/github/ambry/account/DatabaseTest.java
@@ -14,8 +14,8 @@
 package com.github.ambry.account;
 
 import com.github.ambry.account.mysql.ContainerDao;
-import com.github.ambry.account.mysql.MySqlConfig;
 import com.github.ambry.account.mysql.MySqlDataAccessor;
+import com.github.ambry.config.MySqlAccountServiceConfig;
 import com.github.ambry.config.VerifiableProperties;
 import com.github.ambry.utils.Utils;
 import java.sql.Connection;
@@ -54,7 +54,7 @@ public class DatabaseTest {
   }
 
   private static void perfTest(VerifiableProperties verifiableProperties) throws Exception {
-    MySqlConfig config = new MySqlConfig(verifiableProperties);
+    MySqlAccountServiceConfig config = new MySqlAccountServiceConfig(verifiableProperties);
     MySqlDataAccessor dataAccessor = new MySqlDataAccessor(config);
     ContainerDao containerDao = new ContainerDao(dataAccessor);
     // Use high account id to avoid conflict
@@ -91,8 +91,7 @@ public class DatabaseTest {
   private static void cleanup(Connection dbConnection, short startAccountId) throws SQLException {
     Statement statement = dbConnection.createStatement();
     int numDeleted = statement.executeUpdate(
-        "delete from " + ContainerDao.CONTAINER_TABLE + " where " + ContainerDao.ACCOUNT_ID + " >= "
-            + startAccountId);
+        "delete from " + ContainerDao.CONTAINER_TABLE + " where " + ContainerDao.ACCOUNT_ID + " >= " + startAccountId);
     logger.info("Deleted {} containers", numDeleted);
   }
 }
